restructured code and added splash screen
