+++ lib/ui/widgets/custom_app_bar.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+class GradientAppBar extends StatelessWidget {
+  final String title;
+  final double barHeight = 50.0;
+
+  GradientAppBar(this.title);
+
+  @override
+  Widget build(BuildContext context) {
+    final double statusbarHeight = MediaQuery.of(context).padding.top;
+
+    return new Container(
+      height: statusbarHeight + barHeight,
+      width: MediaQuery.of(context).size.width,
+      child: new Center(
+        child: new Text(
+          title,
+          style: new TextStyle(
+              fontSize: 20.0, color: Colors.white, fontWeight: FontWeight.bold),
+        ),
+      ),
+      decoration: new BoxDecoration(
+        gradient: new LinearGradient(
+            colors: [Colors.deepOrange, Colors.pink],
+            begin: const FractionalOffset(0.0, 0.0),
+            end: const FractionalOffset(1.0, 0.0),
+            stops: [0.0, 1.0],
+            tileMode: TileMode.clamp),
+      ),
+    );
+  }
+}
